backend/requirements.txt
Flask==3.0.0
pymongo==4.7.0
dnspython==2.6.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.1
